Reorganizes items so that they are in their own directories and changes inherited variables from "private" to "protected"
