+++ app/src/main/java/com/uncc/gamejam49er/espionage/NewPlayerActivity.java
@@ -2,8 +2,6 @@ package com.uncc.gamejam49er.espionage;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
@@ -11,9 +9,9 @@ import android.widget.Button;
 import android.widget.ImageView;
 
 import com.uncc.gamejam49er.espionage.items.Item;
-import com.uncc.gamejam49er.espionage.items.Shiv;
-import com.uncc.gamejam49er.espionage.items.SmokeBomb;
-import com.uncc.gamejam49er.espionage.items.TwoPieceSuit;
+import com.uncc.gamejam49er.espionage.items.weapons.Shiv;
+import com.uncc.gamejam49er.espionage.items.stats.SmokeBomb;
+import com.uncc.gamejam49er.espionage.items.armor.TwoPieceSuit;
 
 public class NewPlayerActivity extends AppCompatActivity {
     public final static String EXTRA_VIEW_ITEM = "com.uncc.gamejam49er.espionage.EXTRA_VIEW_ITEM";
