--- app/src/main/java/com/uncc/gamejam49er/espionage/items/Stat.java
@@ -1,9 +1,0 @@
-package com.uncc.gamejam49er.espionage.items;
-
-/**
- * Created by Megan on 10/24/2015.
- */
-public class Stat extends Item{
-    public Stat() {
-    }
-}
