+++ app/src/main/java/com/uncc/gamejam49er/espionage/items/armor/AdaptiveCamoflage.java
@@ -1,4 +1,8 @@
-package com.uncc.gamejam49er.espionage.items;
+package com.uncc.gamejam49er.espionage.items.armor;
+
+import com.uncc.gamejam49er.espionage.items.weapons.BroachFlameThrower;
+import com.uncc.gamejam49er.espionage.items.Item;
+import com.uncc.gamejam49er.espionage.items.weapons.KillerNanoBugs;
 
 /**
  * Created by Megan on 10/24/2015.
