+++ app/src/main/java/com/uncc/gamejam49er/espionage/items/stats/SmokeBomb.java
@@ -1,4 +1,4 @@
-package com.uncc.gamejam49er.espionage.items;
+package com.uncc.gamejam49er.espionage.items.stats;
 
 /**
  * Created by Megan on 10/24/2015.
