+++ app/src/main/java/com/uncc/gamejam49er/espionage/items/stats/Stat.java
@@ -1,0 +1,11 @@
+package com.uncc.gamejam49er.espionage.items.stats;
+
+import com.uncc.gamejam49er.espionage.items.Item;
+
+/**
+ * Created by Megan on 10/24/2015.
+ */
+public class Stat extends Item {
+    public Stat() {
+    }
+}
